+++ src/Accounts/AccountType.java
@@ -1,4 +1,4 @@
-package Users;
+package Accounts;
 
 public enum AccountType {
