+++ src/Users/Customer.java
@@ -1,5 +1,7 @@
 package Users;
 
+import Accounts.Account;
+
 import java.util.ArrayList;
 
 public class Customer extends User{
